# Optimize API/Statistics Endpoint Performance